[220109] - carousel 분리
